refactor(003): collapse duplicate anchor registries into constants/anchors.ts
Two independent KNOWN_ANCHORS arrays existed in parallel:
  - constants/index.ts had 5 anchors (Bitso, Flutterwave, MyChoice, Tempo,
    Cowrie) with corridors that did not match the operational registry
  - lib/stellar/anchors.ts had 3 anchors (MoneyGram, Cowrie, Anclap) which
    is what all rate-fetching and SEP-1 resolution code actually used

This drift caused the UI anchor count (5) to disagree with what the app
actually queries, and Cowrie's corridors were out of sync between the two.

- New constants/anchors.ts is the single typed export: ANCHORS, KNOWN_ANCHORS,
  CORRIDORS, ANCHOR_HOME_DOMAINS, USDC_ASSET — no functions, pure data
- lib/stellar/anchors.ts re-exports everything from constants/anchors.ts via
  export * and keeps only the SEP-1 resolution helpers (getAnchorById,
  getAnchorsByCorridorId, getCorridorById, isValidCorridorId,
  discoverAnchorsForCorridor) that require network calls / dynamic imports
- constants/index.ts drops its conflicting KNOWN_ANCHORS literal and
  re-exports the canonical set from ./anchors
- components/ui/CorridorSelector.tsx imports CORRIDORS from @/constants/anchors
  (the canonical source) instead of @/lib/stellar/anchors

All existing imports from @/lib/stellar/anchors and @/constants continue to
resolve correctly; no test changes required.
