+++ lib/stellar/anchors.ts
@@ -1,154 +1,16 @@
-import type { Anchor, Corridor, Sep1TomlData, StellarAsset } from '@/types'
-import { USDC_ISSUER } from '../config'
-// ─── USDC asset ───────────────────────────────────────────────────────────────
+// Anchor and corridor data lives in constants/anchors.ts — the single source of truth.
+// This module re-exports that data and adds SEP-1 resolution helpers that belong
+// in lib/stellar (network calls, dynamic imports) rather than in constants.
+export * from '@/constants/anchors'
 
-/** USDC on Stellar mainnet (Circle issuer). */
-export const USDC_ASSET: StellarAsset = {
-  code: 'USDC',
-  issuer: USDC_ISSUER,
-  name: 'USD Coin',
-}
-
-// ─── Anchors ──────────────────────────────────────────────────────────────────
-
-/**
- * MoneyGram — all five corridors (USDC → NGN / KES / GHS / MXN / BRL).
- * Confirmed live SEP-24 anchor. USDC withdraw enabled, max $2,500/tx.
- * Cash pickup at MoneyGram agent locations worldwide.
- */
-const MONEYGRAM: Anchor = {
-  id: 'moneygram',
-  name: 'MoneyGram',
-  homeDomain: 'stellar.moneygram.com',
-  corridors: ['usdc-ngn', 'usdc-kes', 'usdc-ghs', 'usdc-mxn', 'usdc-brl'],
-  assetCode: 'USDC',
-  assetIssuer: USDC_ISSUER,
-}
-
-/**
- * Cowrie Exchange — Nigeria corridor (USDC → NGN).
- * Uses SEP-6 (TRANSFER_SERVER), not SEP-24. Included for comparison.
- */
-const COWRIE: Anchor = {
-  id: 'cowrie',
-  name: 'Cowrie Exchange',
-  homeDomain: 'cowrie.exchange',
-  corridors: ['usdc-ngn'],
-  assetCode: 'USDC',
-  assetIssuer: USDC_ISSUER,
-}
-
-/**
- * Flutterwave - Nigeria, Kenya, and Ghana corridors (USDC -> NGN / KES / GHS).
- */
-const FLUTTERWAVE: Anchor = {
-  id: 'flutterwave',
-  name: 'Flutterwave',
-  homeDomain: 'flutterwave.com',
-  corridors: ['usdc-ngn', 'usdc-kes', 'usdc-ghs'],
-  assetCode: 'USDC',
-  assetIssuer: USDC_ISSUER,
-}
-
-/**
- * Anclap — Argentina and Peru corridors (USDC → ARS / PEN).
- * Confirmed live SEP-24 anchor. 2% withdrawal fee.
- */
-const ANCLAP: Anchor = {
-  id: 'anclap',
-  name: 'Anclap',
-  homeDomain: 'anclap.com',
-  corridors: ['usdc-ars', 'usdc-pen'],
-  assetCode: 'USDC',
-  assetIssuer: USDC_ISSUER,
-}
-
-/** All supported anchors. */
-export const KNOWN_ANCHORS: Anchor[] = [MONEYGRAM, COWRIE, FLUTTERWAVE, ANCLAP] as const
-export const ANCHORS = KNOWN_ANCHORS
+import { ANCHORS, CORRIDORS } from '@/constants/anchors'
+import type { Anchor, Corridor, Sep1TomlData } from '@/types'
 
 export interface DiscoveredAnchor extends Anchor {
   sep1: Sep1TomlData
   transferServerSep24: string
   webAuthEndpoint: string
 }
-
-/** Maps anchor ID → home domain for quick lookup during SEP-1 resolution. */
-export const ANCHOR_HOME_DOMAINS: Record<string, string> = {
-  moneygram: 'stellar.moneygram.com',
-  cowrie: 'cowrie.exchange',
-  flutterwave: 'flutterwave.com',
-  anclap: 'anclap.com',
-} as const
-
-// ─── Corridors ────────────────────────────────────────────────────────────────
-
-const CORRIDOR_NGN: Corridor = {
-  id: 'usdc-ngn',
-  from: 'USDC',
-  to: 'NGN',
-  countryCode: 'NG',
-  countryName: 'Nigeria',
-}
-
-const CORRIDOR_KES: Corridor = {
-  id: 'usdc-kes',
-  from: 'USDC',
-  to: 'KES',
-  countryCode: 'KE',
-  countryName: 'Kenya',
-}
-
-const CORRIDOR_GHS: Corridor = {
-  id: 'usdc-ghs',
-  from: 'USDC',
-  to: 'GHS',
-  countryCode: 'GH',
-  countryName: 'Ghana',
-}
-
-const CORRIDOR_MXN: Corridor = {
-  id: 'usdc-mxn',
-  from: 'USDC',
-  to: 'MXN',
-  countryCode: 'MX',
-  countryName: 'Mexico',
-}
-
-const CORRIDOR_BRL: Corridor = {
-  id: 'usdc-brl',
-  from: 'USDC',
-  to: 'BRL',
-  countryCode: 'BR',
-  countryName: 'Brazil',
-}
-
-const CORRIDOR_ARS: Corridor = {
-  id: 'usdc-ars',
-  from: 'USDC',
-  to: 'ARS',
-  countryCode: 'AR',
-  countryName: 'Argentina',
-}
-
-const CORRIDOR_PEN: Corridor = {
-  id: 'usdc-pen',
-  from: 'USDC',
-  to: 'PEN',
-  countryCode: 'PE',
-  countryName: 'Peru',
-}
-
-/** All supported corridors. */
-export const CORRIDORS: Corridor[] = [
-  CORRIDOR_NGN,
-  CORRIDOR_KES,
-  CORRIDOR_GHS,
-  CORRIDOR_MXN,
-  CORRIDOR_BRL,
-  CORRIDOR_ARS,
-  CORRIDOR_PEN,
-] as const
 
 // ─── Lookup helpers ───────────────────────────────────────────────────────────
 
@@ -157,10 +19,10 @@ export const CORRIDORS: Corridor[] = [
  * Throws a descriptive error if the ID is not found.
  */
 export function getAnchorById(id: string): Anchor {
-  const anchor = KNOWN_ANCHORS.find((a) => a.id === id)
+  const anchor = ANCHORS.find((a) => a.id === id)
   if (!anchor) {
     throw new Error(
-      `Unknown anchor: "${id}". Valid IDs: ${KNOWN_ANCHORS.map((a) => a.id).join(', ')}`
+      `Unknown anchor: "${id}". Valid IDs: ${ANCHORS.map((a) => a.id).join(', ')}`
     )
   }
   return anchor
@@ -171,7 +33,7 @@ export function getAnchorById(id: string): Anchor {
  * Returns an empty array if no anchors support the corridor.
  */
 export function getAnchorsByCorridorId(corridorId: string): Anchor[] {
-  return KNOWN_ANCHORS.filter((a) => a.corridors.includes(corridorId))
+  return ANCHORS.filter((a) => a.corridors.includes(corridorId))
 }
 
 /**
@@ -180,12 +42,16 @@ export function getAnchorsByCorridorId(corridorId: string): Anchor[] {
  */
 export async function discoverAnchorsForCorridor(corridorId: string): Promise<DiscoveredAnchor[]> {
   const { resolveToml } = await import('./sep1')
-  const corridorAnchors = KNOWN_ANCHORS.filter((anchor) => anchor.corridors.includes(corridorId))
+  const corridorAnchors = ANCHORS.filter((anchor) => anchor.corridors.includes(corridorId))
 
   const results = await Promise.allSettled(
     corridorAnchors.map(async (anchor): Promise<DiscoveredAnchor> => {
-      const sep1 = await resolveToml(anchor.homeDomain)
-
+      const result = await resolveToml(anchor.homeDomain)
+      if (!result.ok) throw new Error(result.error)
+      const sep1 = result.data
+      if (!sep1.TRANSFER_SERVER_SEP0024 || !sep1.WEB_AUTH_ENDPOINT) {
+        throw new Error(`Anchor "${anchor.id}" does not support SEP-24 or SEP-10.`)
+      }
       return {
         ...anchor,
         sep1,
